Bumped all integer tags to int64 to avoid data loss

+++ Source/PCGExtendedToolkit/Private/Graph/PCGExFindClustersData.cpp
@@ -64,7 +64,7 @@ bool FPCGExFindClustersDataElement::Boot(FPCGExContext* InContext) const
 			}
 		}
 
-		Context->SearchKey = Context->SearchKeyIO->Tags->GetTypedValue<int32>(PCGExGraph::TagStr_PCGExCluster);
+		Context->SearchKey = Context->SearchKeyIO->Tags->GetTypedValue<int64>(PCGExGraph::TagStr_PCGExCluster);
 		if (!Context->SearchKey)
 		{
 			PCGE_LOG_C(Error, GraphAndLog, InContext, FTEXT("Found no valid key to match against."));
